+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "pallet-ismp/runtime-api",
     "pallet-ismp/primitives",
     "pallet-ismp/primitives/state-machine",
+    "pallet-ismp/evm",
     "pallet-ismp",
     "parachain/inherent",
     "parachain/runtime-api",
