+++ pallet-ismp/Cargo.toml
@@ -18,10 +18,6 @@ pallet-timestamp = { git = "https://github.com/paritytech/substrate", branch = "
 
 # EVM support
 pallet-evm = { git = "https://github.com/paritytech/frontier", branch = "polkadot-v0.9.42", default-features = false }
-fp-evm = { git = "https://github.com/paritytech/frontier", branch = "polkadot-v0.9.42", default-features = false, optional = true }
-alloy-sol-macro = { version = "0.2.0", optional = true }
-alloy-sol-types = { version = "0.2.0", default-features = false, optional = true }
-alloy-primitives = { version = "0.2.0", default-features = false, optional = true }
 
  # polytope labs
 ismp-rs = { package = "ismp", git = "https://github.com/polytope-labs/ismp-rs", branch = "main", default-features = false }
@@ -63,7 +59,6 @@ std = [
     "ismp-primitives/std",
     "pallet-evm/std"
 ]
-evm = ["fp-evm", "alloy-primitives", "alloy-sol-types", "alloy-sol-macro"]
 
 
 runtime-benchmarks = [
