+++ pallet-ismp/evm/Cargo.toml
@@ -1,0 +1,56 @@
+[package]
+name = "ismp-evm"
+version = "0.1.0"
+edition = "2021"
+authors = ["Polytope Labs <hello@polytope.technology>"]
+
+[dependencies]
+# substrate
+frame-support = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false }
+frame-system = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false }
+sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false }
+sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false }
+sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false }
+sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false }
+
+# EVM support
+pallet-evm = { git = "https://github.com/paritytech/frontier", branch = "polkadot-v0.9.42", default-features = false }
+fp-evm = { git = "https://github.com/paritytech/frontier", branch = "polkadot-v0.9.42", default-features = false }
+alloy-sol-macro = { version = "0.2.0" }
+alloy-sol-types = { version = "0.2.0", default-features = false }
+alloy-primitives = { version = "0.2.0", default-features = false }
+
+ # polytope labs
+ismp-rs = { package = "ismp", git = "https://github.com/polytope-labs/ismp-rs", branch = "main", default-features = false }
+
+# crates.io
+codec = { package = "parity-scale-codec", version = "3.1.3", default-features = false }
+scale-info = { version = "2.1.1", default-features = false, features = ["derive"] }
+
+# local
+ismp-primitives = { path = "../primitives", default-features = false }
+pallet-ismp = { path = "../", default-features = false }
+
+[dev-dependencies]
+
+
+[features]
+default = ["std"]
+std = [
+    "codec/std",
+    "frame-support/std",
+    "frame-system/std",
+    "scale-info/std",
+    "sp-io/std",
+    "sp-runtime/std",
+    "sp-std/std",
+    "sp-core/std",
+    "ismp-rs/std",
+    "ismp-primitives/std",
+    "pallet-evm/std",
+    "fp-evm/std",
+    "alloy-primitives/std",
+    "alloy-sol-types/std",
+    "pallet-ismp/std"
+]
+
