--- pallet-ismp/src/evm/mod.rs
@@ -1,6 +1,0 @@
-//! EVM contract support
-
-pub mod abi;
-pub mod handler;
-pub mod ismp_dispatcher_precompiles;
-pub mod weight;
