+++ pallet-ismp/src/mocks/ismp.rs
@@ -7,7 +7,6 @@ use ismp_rs::{
         ConsensusClient, StateCommitment, StateMachineClient, StateMachineHeight, StateMachineId,
         VerifiedCommitments,
     },
-    contracts::Gas,
     error::Error as IsmpError,
     handlers,
     host::{Ethereum, IsmpHost, StateMachine},
@@ -32,16 +31,16 @@ where
 pub struct MockModule;
 
 impl IsmpModule for MockModule {
-    fn on_accept(&self, _request: Post) -> Result<Gas, ismp_rs::error::Error> {
-        Ok(().into())
+    fn on_accept(&self, _request: Post) -> Result<(), ismp_rs::error::Error> {
+        Ok(())
     }
 
-    fn on_response(&self, _response: Response) -> Result<Gas, ismp_rs::error::Error> {
-        Ok(().into())
+    fn on_response(&self, _response: Response) -> Result<(), ismp_rs::error::Error> {
+        Ok(())
     }
 
-    fn on_timeout(&self, _request: Request) -> Result<Gas, ismp_rs::error::Error> {
-        Ok(().into())
+    fn on_timeout(&self, _request: Request) -> Result<(), ismp_rs::error::Error> {
+        Ok(())
     }
 }
